Fix for clean build on both 32 and 64 bit systems
Revision to commit 38b291c50441b77

+++ src/common/slurm_protocol_pack.c
@@ -2401,16 +2401,18 @@ _unpack_network_callerid_msg(network_callerid_msg_t **msg_ptr, Buf buffer,
 	if (protocol_version >= SLURM_MIN_PROTOCOL_VERSION) {
 		safe_unpackmem_xmalloc(&charptr_tmp, &uint32_tmp, buffer);
 		if (uint32_tmp > (uint32_t)sizeof(msg->ip_src)) {
-			error("%s: ip_src that came across is %u and we can only handle %u",
-			      __func__, uint32_tmp, sizeof(msg->ip_src));
+			error("%s: ip_src that came across is %u and we can only handle %lu",
+			      __func__, uint32_tmp,
+			      (long unsigned) sizeof(msg->ip_src));
 			goto unpack_error;
 		}
 		memcpy(msg->ip_src, charptr_tmp, (size_t)uint32_tmp);
 		xfree(charptr_tmp);
 		safe_unpackmem_xmalloc(&charptr_tmp, &uint32_tmp, buffer);
 		if (uint32_tmp > (uint32_t)sizeof(msg->ip_dst)) {
-			error("%s: ip_dst that came across is %u and we can only handle %u",
-			      __func__, uint32_tmp, sizeof(msg->ip_dst));
+			error("%s: ip_dst that came across is %u and we can only handle %lu",
+			      __func__, uint32_tmp,
+			      (long unsigned) sizeof(msg->ip_dst));
 			goto unpack_error;
 		}
 		memcpy(msg->ip_dst, charptr_tmp, (size_t)uint32_tmp);
